// modules/dto/register.dto.ts
class Register {
  id?: number;
  name?: string;
  actcode?: string;
  description?: string;
  standard?: string;
  duratypeid?: number;
  is_active?: boolean;
  machinedetailsid?: string;
  result_type?: number;
  machinegroupid?: number;
  checkyear?: number;
  monthcurrent?: number;
  season_id?: number;
  update_date?: Date | null;
  update_by?: string;
}

export default Register;